* modules/em8300_main.c 	merge patch from Frédéric Crozat, 	"A cast is needed in switch, otherwise em8300.o 	needs a function which is only present in 	libgcc.a (compiler library) :  __cmpdi2"
	* modules/adv717x.c
	added swap_redblue_pal for people with
	swapped red and blue components;
	changes Timing Register 1 for PAL and PAL60.
	register values from Frédéric Crozat.

+++ modules/em8300_main.c
@@ -320,7 +320,7 @@ int em8300_io_mmap(struct file *file, struct vm_area_struct *vma)
 	return -EPERM;
 
 #if LINUX_VERSION_CODE < KERNEL_VERSION(2,3,0)
-    switch(vma->vm_offset) {
+    switch((unsigned long) vma->vm_offset) {
 #else	
     switch(vma->vm_pgoff) {
 #endif
